Rename --verify for server resize to avoid conflict
Currently server resize command has a verify argument for user to
confirm resize operation. But this argument is also registered by
openstack client shell which is used to enable server certificate
verification. So verify argument can not be correctly passed to
server resize command.

This patch renames --verify to --confirm (nova client has a
resize-confirm command) to solve this problem.

Change-Id: I751aa3df16c3b637b82944dd350afc5cbab62120
Closes-Bug: #1416542