Print warning on authentication error
At least print a warning on authentication error.  I have no
idea why an exception is being ignored here because if there
is no session, nothing is going to happen.  This at least will
print some useful warning:

    (.venv)terry@f350:~/python-openstackclient$ os flavor list --os-cloud pro
    WARNING: openstackclient.shell Possible error authenticating: __init__() got an unexpected keyword argument 'asdf'
    ERROR: openstack Authentication requires 'auth_url', which should be specified in 'HTTPClient'

This error was caused by having a bogus value 'asdf' in the cloud.yaml
for the cloud pro.

Change-Id: Ie08432e0464cfa86b3b3f67ca29d3b7d23d2f46f