Glance `image set` Resolve Fracturing
Currently `image set` uses the new api, where other parts of osc the old
api is used. This deprecates the v2 api in favour of the v1 to maintain
the same commands across osc. However the functionality now remains
there as people could now be using this functionality.

This also adds the --unprotected argument, as in the previous version if
--protected was not supplied it would just make the argument
--unprotected without the users explicit consent.

The patch also fixes the documentation for image set as it was outdated.

Change-Id: I990d20332c80165102badef7ac94ddbeb7824950
Closes-Bug: 1498092