Set default network api to 2.0 instead of 2
neutronclient expects 2.0 as the version if you go through the discovery
constructor. For that reason, 2.0 is the 'correct' version to set in
config files or environment variables for if you're using things that
are not OSC. However, if you do that, OSC prints a warning that 2.0 is
not in the supported network version list.

Let's support both so that users don't get a confuse.

Change-Id: I7412519693f75fcd29f5621ce9e5a2df2da92684