Add class TestServerList to provide basic unit test for "server list" command.
This patch provide a class to test "server list" command. Only one simplest
case in this patch.

Some of the options in "server list" are complicated. And the server object
contains lots of attributes need to be handled in specific ways. So other
test cases will be added in other patches.

Change-Id: Id9fdba8f149bd74187aa42516067dacebc6962b5
Implements: blueprint osc-unit-test-framework-improvement