Fix showing network quotas for a project
The OpenStack SDK is now used for the network client. However,
the 'openstack quota show' command wasn't updated for the
client change. As a result, the command will fail to show
network quotas when a project name is specified. For example:

  $ openstack quota show admin
  'Proxy' object has no attribute 'show_quota'

This patch set fixes the command by using the OpenStack SDK
to get network quotas for a project.

Change-Id: I59a7b6780a7b80cd09e79d40d214751b25d3016e
Related-To: blueprint neutron-client
Closes-Bug: #1528249