Refactor: Set "project_id" for FakeXXX in a consistent style
OpenStack SDK will translate "project_id" into "tenant_id" automatically
when referring to "tenant_id" attribute with the name "project_id". So
when faking an object returned fron SDK, we need to fake this behavior.

The original way is ugly. This patch turns it into a consistent style,
and give better comments.

Change-Id: I0dfb1f7552fc28eb4e7ebf5c614c9f3bde79ad80