Fix functional test failures
There have been a lot of functional test failures for compute
and network.  I believe they are timing related in that the
test_server is sometimes booting a server with a flavor,
image and/or network created by another functional test. Such
resources are then deleted by the owning functional tests
causing various types of failures

Change-Id: Ib82edd487da46610374fb0eb5796cb98f2d3aa2c