Set up 3.x release notes
Give the 2.x series its own release notes page, put the 3.x notes into
the 'current' series page.  This removes the OpenStack development cycle
name from the release notes.

We follow the intermediary release model, however the association with
the common cycle release is something that we do not need to reflect in
the doc structure.  We want users to use the current version even with
older clouds, not to think they need to use a 'Mitaka' OSC with a Mitaka
cloud.

Change-Id: Idbb6844824d50e5cd0fa64871b5aa625d4d237af