Change to plural form of object in multi delete error message in networkv2
Usually, the error message of multi delete is:
"'result' of 'total' 'objects' failed to delete"
the objects is a plural form.
To match the other multi delete error messages
in OSC, change the object in delete error message
in networkv2 to a plural form. Just add a 's' in
the message.

Change-Id: I17e0735d025bb61014db709d2639813565015b3d