Mask passwords in debug logs for auth_config_hook
The auth config hook can have credentials in it so
we have to mask the config before logging it. To
avoid doing the work of masking the password if we
aren't going to log it, there is a conditional put
around the actual debug statement.

Conflicts:
        openstackclient/common/client_config.py

NOTE(mriedem): The conflict was due to imports.

Change-Id: I8e626672ec94fc837610216bccb4354dbdedca17
Closes-Bug: #1630822
(cherry picked from commit cd1a412408f068aeef97c1ee368400307fce7733)