Beef up address scope functional tests
We need to get more thorough in our functional testing, so start by
adding tests for all create and set options, check return values.

This also removes most of the setupClass() and teardownClass()
methods as they held common state that was subject to race conditions
when running tests in parallel.

Change-Id: Ib337f9e9d16b4183bb319b58cbe943045f365ff2