Use image client for images instead of compute
With the deprecation of the Nova proxy APIs in microversion 2.36 [1],
any operation that uses a microversion higher than 2.36 and works with
images will fail because the /images endpoint will return 404. This
patch updates openstackclient to query images using the image client
in places where previously the compute client was used.

[1] http://docs.openstack.org/developer/nova/api_microversion_history.html#id33

Change-Id: Ia66e44e530799ce6531922dcf6a84e38528c8725
Closes-bug: 1630161