Use *_as_ids instead *_as_list
The parents_as_list and subtree_as_list query parameters limit the
result to only parents and subtree where the user making the call has
role assignments in. Since OSC only displays the IDs, the call would be
the same as the similar *_as_ids queries, the difference is that the
later doesn't enforce the role assignments (making it more useful).

Output example by using this patch:

$ openstack project show --children root
+-------------+------------------------------+
| Field       | Value                        |
+-------------+------------------------------+
| description |                              |
| domain_id   | default                      |
| enabled     | True                         |
| id          | 123                          |
| is_domain   | False                        |
| name        | root                         |
| parent_id   | default                      |
| subtree     | {u'456': None, u'789': None} |
+-------------+------------------------------+

Change-Id: Ib7b37ae8f55190a7efcc375d5be4a2823d02d1a4