Simplify logic around option lists in port set
Use a common pattern to handle option pairs --XYZ and --no-XYZ for managing
lists of attributes. This pattern looks at the presence of the option
in parsed_args first and branches as necessary.

Some specific steps are included for the SDK Network resources to reliably
set the 'dirty' flag for changed attributes via one or both of the following:
* iterate over lists of original resource attributes to force the creation
  of a new list object
* use [].extend() rather than += to add to the existing list (substitute
  {}.update() for dicts)

Change-Id: I0c3f9a52ffe1ae2b5b230cb13d6376dd9131aaf9