doc: Remove local fork of apidoc
This is unnecessary as pbr has since been fixed. It was causing a broken
build as it didn't respect the '[pbr] autodoc_tree_excludes' setting in
setup.cfg.

The 'openstackclient/volume/v3' directory is an empty module containing
only an '__init__' file). Empty modules = unhappy autodoc, thus, this
module is ignored.

Change-Id: Ie355b14c14b7badccb5c25f7c17edbc5e4b3804f