Use openstackdocstheme in release note
In the release notes build, openstackdocstheme is specified
in the sphinx extension list, but it is actually not used.
This commit configures openstackdocstheme as suggested in
the openstackdocstheme document.

Also specifies display_toc False in html_theme_options.
This prevents openstackdocstheme to add the local TOC automatically.
OSC has several releases during a single development cycle and
it leads to a long TOC at the top of individual pages.
It loses the readability and we can see version numbers in the
left sidebar, so we can safely disable the local TOC in
the release notes.

[1] https://docs.openstack.org/openstackdocstheme/latest/

Change-Id: Iae08d309f7589bb13a1766f6fded70673ba24047