Added AddNetwork command to server
Currently, if users want to add another NIC to a running instance,
they need to (i) create a neutron port and (ii) add the port to the
server via teh AddPort command. It would be more convenient to have
a single command to achieve the equivalent.

Novaclient already support adding network to an instance via the
interface-attach command. This patch introduces a similar capability
in OSC.

Change-Id: Ia3e39c57ae7ecb96aae1b66adc52c289daccb6ec