Attempt to de-race qos policy
We're getting about 1-in-6 failures on qos policy delete now, with the message
that the policy is in use by a network.  It shouldn't be, this is possibly
due to the small window where the policy is set as the default.  Let's
remove that and shore up the test using --share instead.

Change-Id: I8d669bd3c5c88dadd2927aee89e5ef72cf4001c4