Make py27 and py35 versions and template of unit-tips jobs
Make py27 and py35 versions of the tips job, and a project-template to
hold them so it's easy for other projects to use both (or more, we ever
add them) The base tox job knows how to install siblings based on
required-projects, which makes it easy to piggyback on them.

Remove the irrelevant-files sections, as these are specified in the
base openstack-tox jobs.

Using the unit-tips or functional-tips tox env is no longer neccessary.
For the gate it's actually important to not do the sibling processing by
hand in tox as the base tox job version of the logic does all the
right things to deal with constraints and whatnot. Leave them for local
developer convenience.

Leave the osc-tox-unit-tips job for now, since there are some other
repos using it. Once we switch them to the template, we can remove the
job.

Change-Id: I599b18218c10cb08e508cca3b3bbc9c88b8f809c