Allow ports filtering with device_id
Right now, if a neutron port is owned by a container powered by
Kuryr, there is no way to list and filter those ports because
OSC assumed a neutron port is owned by either a server or router.
This patch adds support for that by introducing an option '--device-id'
to the 'port list' command.

Change-Id: Ib1fd27e8d843a99fb02ccabd8a12a24ac27cec9c